requirements.txt
PyQt6
yt-dlp
ytmusicapi
google-api-python-client
google-auth-oauthlib
qdarkstyle
